+++ test/stub/shell.js
@@ -3,7 +3,7 @@ const Shell = require('../../lib/shell');
 
 module.exports = class ShellStub extends Shell {
   exec(command) {
-    if (/^(npm|git fetch)/.test(command)) {
+    if (/^(npm (ping|publish|show|whoami)|git fetch)/.test(command)) {
       debug(command);
       return Promise.resolve();
     }